pyproject.toml
[project]
name = "langrove"
version = "0.1.0"
description = "Open-source, self-hosted drop-in replacement for LangGraph deployment. Agent Protocol compliant."
readme = "README.md"
authors = [
    { name = "Devang Padhiyar", email = "devangpadhiyar700@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "asyncpg>=0.30",
    "redis[hiredis]>=5.0",
    "orjson>=3.10",
    "langgraph>=0.6",
    "langgraph-checkpoint-postgres>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "click>=8.0",
    "python-dotenv>=1.0",
    "alembic>=1.14",
    "psycopg>=3.3.3",
    "sqlalchemy>=2.0",
    "langchain-openai>=1.1.12",
    "langchain>=1.2.14",
    "psycopg-pool>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "testcontainers[postgres,redis]>=4.0",
    "coverage>=7.0",
]

[project.scripts]
langrove = "langrove.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/langrove"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
